requirements.txt
Flask
python-dotenv
pdfminer.six
python-docx
requests
